CLJS-414: specify and specify* macros
These allow an instance to directly implement a protocol.

specify has a syntax similar to extend-type and constructs its method closures.
specify* is lower level with a syntax loosely based on clojure.core/extend. It takes a closure for every specified protocol-method-arity point. This allows to reuse the closures, reducing allocation.